# tracer: add option to skip and limit stack frames