--- a/ddtrace/tracer/tracer_test.go
+++ b/ddtrace/tracer/tracer_test.go
@@ -1074,21 +1074,33 @@ func cpspan(s *span) *span {
 	}
 }
 
-func Test_takeStackTrace(t *testing.T) {
-	t.Run("skips self", func(t *testing.T) {
+func TestTakeStackTrace(t *testing.T) {
+	t.Run("n=12", func(t *testing.T) {
 		val := takeStacktrace(12, 0)
 		// top frame should be runtime.main or runtime.goexit, in case of tests that's goexit
-		assert.Contains(t, val, "runtime.goexit", "should contain ")
-		assert.Contains(t, val, "testing.tRunner", "should contain parent function")
-		assert.Contains(t, val, "tracer.Test_takeStackTrace", "should contain this function")
+		assert.Contains(t, val, "runtime.goexit")
+		assert.Contains(t, val, "testing.tRunner")
+		assert.Contains(t, val, "tracer.TestTakeStackTrace")
 	})
 
-	t.Run("takes requested number of frames", func(t *testing.T) {
+	t.Run("n=15,skip=2", func(t *testing.T) {
+		val := takeStacktrace(3, 2)
+		// top frame should be runtime.main or runtime.goexit, in case of tests that's goexit
+		assert.Contains(t, val, "runtime.goexit")
+		numFrames := strings.Count(val, "\n\t")
+		assert.Equal(t, 1, numFrames)
+	})
+
+	t.Run("n=1", func(t *testing.T) {
 		val := takeStacktrace(1, 0)
-		assert.Contains(t, val, "tracer.Test_takeStackTrace", "should contain this function")
+		assert.Contains(t, val, "tracer.TestTakeStackTrace", "should contain this function")
 		// each frame consists of two strings separated by \n\t, thus number of frames == number of \n\t
 		numFrames := strings.Count(val, "\n\t")
-		assert.Equal(t, 1, numFrames, "should have length one if only one frame was requested")
+		assert.Equal(t, 1, numFrames)
+	})
+
+	t.Run("invalid", func(t *testing.T) {
+		assert.Empty(t, takeStacktrace(100, 115))
 	})
 }