Adds option to limit the number of reported stack frames
as well as skip requested number of stack frames.

Fixes: https://github.com/DataDog/dd-trace-go/issues/435