@matoous 2019-05-14
> Thanks. We'll need to think about this a bit more to make sure we're doing it right. I left some comments. We'd also need to have some benchmarks. It would be nice to compare this to the regular `debug.Stack` call using something like [`benchcmp`](https://godoc.org/golang.org/x/tools/cmd/benchcmp)

I added benchmarks, the speed is on par with `debug.stack()`, both averaging ~ 1088 ns/op.