@matoous 2019-05-15
Benchmark of new implementation:
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkTracerStackFrames-8     1049          1050          +0.10%
```