@matoous 2019-05-15
@gbbr 
> Lastly, can you please submit additional commits for each review round. It would be nice to compare previous versions to newer versions you are making. We can then squash at the very end. Unfortunately with Github it's not possible to go back to previous revisions without having multiple commits in a PR.

I will admit, I don't know how to do that. Can you help me? I am used to ammend all the commits because that's what we do in our team and I forgot not to do it elsewhere.