@matoous 2019-05-18
I hoped that I fixed everything necessary, I am getting lost in the comments 😄