@gbbr 2019-05-20
I added another commit with some improvements to the tests to make sure there's no funny edge-cases.

I also renamed the sub-tests a bit because even though I prefer lengthier descriptions, in practice, they look pretty bad when using the `-v` command. I noticed that in Go it's preferred to have very short (if possible one word) names for sub-tests. This is what you would get without the change:
```
=== RUN   Test_takeStackTrace
=== RUN   Test_takeStackTrace/skips_self
=== RUN   Test_takeStackTrace/takes_requested_number_of_frames
```
Looks a bit weird IMO