@gbbr 2019-05-21
Released as part of https://github.com/DataDog/dd-trace-go/releases/tag/v1.14.0