@mrVanboy 2019-05-23
After rebasing our project to the new version with `StackFrames`, I realized, that it requires changes in many `contrib/*` subpackages too. Our usecase is gRPC middleware for example. 

Maybe it worth time to implement similar functionality as a `tracer.StartOption` to configure frames only once at a single place during tracer initialization? We could prepare another PR for that.