@gbbr 2019-05-23
@mrVanboy sure! Let's discuss how we'll approach that in a separate Github issue. Would you mind opening one? We're of course happy to take suggestions to solutions and contributions. We'll likely need a new [`Option`](https://godoc.org/gopkg.in/DataDog/dd-trace-go.v1/contrib/google.golang.org/grpc#Option).